Revert "Retry: Add missing com.android.media.swcodec to apex_available"
This reverts commit 74d23b5ca9042ee3b3ced4839001fbb226fd2647.

Reason for revert: Not needed

The original change was made in an attempt to close a perceived hole in
the apex availability checking. However, there is no such hole and so
this change is unnecessary. While there is a dependency path from the
apex to this module one of the dependencies in that path is from a
static library to a shared library. That dependency does not result in
the code that includes the static library being linked to the shared
library at either build time or runtime. Therefore, the shared library
is not part of the APEX and so does not need to be marked as such.

Bug: 152762638
Test: m droid
Change-Id: I9b08ead5cfb5fd506d8446118398e844f77eec98